refactor: SRCに移動した中でも、UIに関係ありそうなところは、Presentaitonへと一層落とす。
