+++ src/presentation/components/Logo.vue
