+++ src/presentation/components/ProductSelectorDialog.vue
