+++ src/presentation/components/TwoPainWithSlideBarLayout.vue
