+++ src/presentation/components/VuetifyLogo.vue
