+++ src/presentation/components/diagrams/editor/businesscontextdiagram/CanvasSettingToolBar.vue
