+++ src/presentation/components/diagrams/editor/businesscontextdiagram/draw2d/eventanalyze/FigureAnalyzer.ts
