+++ src/presentation/components/diagrams/editor/businesscontextdiagram/icon/CompanyIconGenerator.ts
