+++ src/presentation/components/diagrams/editor/businesscontextdiagram/icon/RoomIconGenerator.ts
