+++ src/presentation/components/diagrams/editor/statemodel/StateModelEditor.vue
