+++ src/presentation/components/diagrams/editor/systemcontextdiagram/SystemContextDiagramEditor.vue
