+++ src/presentation/components/diagrams/icon/GenericAreaIconGenerator.ts
