+++ src/presentation/components/diagrams/icon/GenericSingleIconGenerator.ts
