+++ src/presentation/components/diagrams/icon/IconGenerator.ts
